git-svn-id: http://openomoikane2.googlecode.com/svn/trunk/Omoikane@36 a36d047a-ba86-11dd-b55b-2da8d274f894

+++ src/omoikane/principal/Grupos.groovy
@@ -14,6 +14,8 @@ import java.awt.*;
 import omoikane.sistema.*;
 import javax.swing.event.*;
 import java.awt.event.*;
+import static omoikane.sistema.Usuarios.*;
+import static omoikane.sistema.Permisos.*;
 
 /**
  *
@@ -26,100 +28,90 @@ class Grupos {
       
     static def lanzarCatalogo()
     {
-        def cat = (new omoikane.formularios.CatalogoGrupos())
-        cat.setVisible(true);
-        escritorio.getPanelEscritorio().add(cat)
-        Herramientas.In2ActionX(cat, KeyEvent.VK_ESCAPE, "cerrar"   ) { cat.btnCerrar.doClick()   }
-        cat.txtBusqueda.keyReleased = { if(it.keyCode == it.VK_ESCAPE) cat.btnCerrar.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_DELETE, "eliminar" ) { cat.btnEliminas.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F4    , "detalles" ) { cat.btnDetalles.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F5    , "nuevo"    ) { cat.btnNuevo.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F6    , "modificar") { cat.btnModificar.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F7    , "imprimir" ) { cat.btnImprimir.doClick() }
-        Herramientas.iconificable(cat)
-        cat.toFront()
-        try { cat.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario catálogo de Grupos", Herramientas.getStackTraceString(e)) }
-        cat.txtBusqueda.requestFocus()
-        poblarGrupos(cat.getTablaGrupos(),"")
+        if(cerrojo(PMA_ABRIRGRUPO)){
+            def cat = (new omoikane.formularios.CatalogoGrupos())
+            cat.setVisible(true);
+            escritorio.getPanelEscritorio().add(cat)
+            Herramientas.In2ActionX(cat, KeyEvent.VK_ESCAPE, "cerrar"   ) { cat.btnCerrar.doClick()   }
+            cat.txtBusqueda.keyReleased = { if(it.keyCode == it.VK_ESCAPE) cat.btnCerrar.doClick() }
+            Herramientas.In2ActionX(cat, KeyEvent.VK_DELETE, "eliminar" ) { cat.btnEliminas.doClick() }
+            Herramientas.In2ActionX(cat, KeyEvent.VK_F4    , "detalles" ) { cat.btnDetalles.doClick() }
+            Herramientas.In2ActionX(cat, KeyEvent.VK_F5    , "nuevo"    ) { cat.btnNuevo.doClick() }
+            Herramientas.In2ActionX(cat, KeyEvent.VK_F6    , "modificar") { cat.btnModificar.doClick() }
+            Herramientas.In2ActionX(cat, KeyEvent.VK_F7    , "imprimir" ) { cat.btnImprimir.doClick() }
+            Herramientas.iconificable(cat)
+            cat.toFront()
+            try { cat.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario catálogo de Grupos", Herramientas.getStackTraceString(e)) }
+            cat.txtBusqueda.requestFocus()
+            poblarGrupos(cat.getTablaGrupos(),"")
+            return cat
+        }else{Dialogos.lanzarAlerta("Acceso Denegado")}
 
     }
 
     static def lanzarCatalogoDialogo()
     {
-        def foco = new Object()
-        def cat = (new omoikane.formularios.CatalogoGrupos())
-        cat.setVisible(true);
-        escritorio.getPanelEscritorio().add(cat)
-        Herramientas.In2ActionX(cat, KeyEvent.VK_ESCAPE, "cerrar"   ) { cat.btnCerrar.doClick()   }
-        cat.txtBusqueda.keyReleased = { if(it.keyCode == it.VK_ESCAPE) cat.btnCerrar.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_DELETE, "eliminar" ) { cat.btnEliminas.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F4    , "detalles" ) { cat.btnDetalles.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F5    , "nuevo"    ) { cat.btnNuevo.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F6    , "modificar") { cat.btnModificar.doClick() }
-        Herramientas.In2ActionX(cat, KeyEvent.VK_F7    , "imprimir" ) { cat.btnImprimir.doClick() }
-        Herramientas.iconificable(cat)
-        cat.toFront()
-        try { cat.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario catalogo de grupos", Herramientas.getStackTraceString(e)) }
-        cat.txtBusqueda.requestFocus()
+        def foco=new Object()
+        def cat = lanzarCatalogo()
+        cat.setModoDialogo()
         cat.internalFrameClosed = {synchronized(foco){foco.notifyAll()} }
         cat.txtBusqueda.keyPressed = { if(it.keyCode == it.VK_ENTER) cat.btnAceptar.doClick() }
         def retorno
-        cat.btnAceptar.actionPerformed = { def catTab = cat.tablaGrupos; retorno = catTab.getModel().getValueAt(catTab.getSelectedRow(), 0) as int; cat.btnCerrar.doClick(); }
-        poblarGrupos(cat.getTablaGrupos(),"")
+        cat.btnAceptar.actionPerformed = { def catTab = cat.tablaGrupos; retorno = catTab.getModel().getValueAt(catTab.getSelectedRow(), 0) as String; cat.btnCerrar.doClick(); }
         synchronized(foco){foco.wait()}
         retorno
     }
 
     static def lanzarFormNuevoGrupo()
     {
-        def formGrupo = new omoikane.formularios.Grupo()
-        formGrupo.setVisible(true)
-        escritorio.getPanelEscritorio().add(formGrupo)
-        Herramientas.iconificable(formGrupo)
-        formGrupo.toFront()
-        try { formGrupo.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario detalles Grupo", Herramientas.getStackTraceString(e)) }
-        formGrupo.setEditable(true);
-        formGrupo.setModoNuevo();
-        formGrupo
+        if(cerrojo(PMA_MODIFICARGRUPO)){
+            def formGrupo = new omoikane.formularios.Grupo()
+            formGrupo.setVisible(true)
+            escritorio.getPanelEscritorio().add(formGrupo)
+            Herramientas.iconificable(formGrupo)
+            formGrupo.toFront()
+            try { formGrupo.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario detalles Grupo", Herramientas.getStackTraceString(e)) }
+            formGrupo.setEditable(true);
+            formGrupo.setModoNuevo();
+            return formGrupo
+        }else{Dialogos.lanzarAlerta("Acceso Denegado")}
     }
 
-        static def guardar(formGrupo)
+    static def guardar(formGrupo)
     {
-        Herramientas.verificaCampos{
-        def descripcion = formGrupo.getTxtDescripcion()
-        def descuento   = formGrupo.getTxtDescuento()
-
-        Herramientas.verificaCampo(formGrupo.getTxtDescripcion(),/^([a-zA-Z0-9_\-\s\ñ\Ñ\*\+áéíóúü]+)$/,"Descripcion sólo puede incluír números, letras, espacios, á, é, í, ó, ú, ü, _, -, * y +.")
-        Herramientas.verificaCampo(formGrupo.getTxtDescuento(),/^([0-9]*[\.]{0,1}[0-9]+)$/,"Descuento sólo puede incluír números reales positivos")
-
-        descuento = descuento as Double
-
-        try {
-            def serv = Nadesico.conectar()
-            Dialogos.lanzarAlerta(serv.addGrupo(descripcion, descuento))
-        } catch(e) { Dialogos.error("Error al enviar a la base de datos. El grupo no se registró", e) }
-
-        formGrupo.dispose()
-        }
+        if(cerrojo(PMA_MODIFICARGRUPO)){
+            Herramientas.verificaCampos{
+                def descripcion = formGrupo.getTxtDescripcion()
+                def descuento   = formGrupo.getTxtDescuento()
+                Herramientas.verificaCampo(formGrupo.getTxtDescripcion(),/^([a-zA-Z0-9_\-\s\ñ\Ñ\*\+áéíóúü]+)$/,"Descripcion sólo puede incluír números, letras, espacios, á, é, í, ó, ú, ü, _, -, * y +.")
+                Herramientas.verificaCampo(formGrupo.getTxtDescuento(),/^([0-9]*[\.]{0,1}[0-9]+)$/,"Descuento sólo puede incluír números reales positivos")
+                descuento = descuento as Double
+                try {
+                    def serv = Nadesico.conectar()
+                    Dialogos.lanzarAlerta(serv.addGrupo(descripcion, descuento))
+                } catch(e) { Dialogos.error("Error al enviar a la base de datos. El grupo no se registró", e) }
+                formGrupo.dispose()
+            }
+        }else{Dialogos.lanzarAlerta("Acceso Denegado")}
     }
 
     static def lanzarDetallesGrupo(ID)
     {
-        def formGrupo = new omoikane.formularios.Grupo()
-        formGrupo.setVisible(true)
-        escritorio.getPanelEscritorio().add(formGrupo)
-        Herramientas.iconificable(formGrupo)
-        formGrupo.toFront()
-        try { formGrupo.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario detalles Grupo", Herramientas.getStackTraceString(e)) }
-
-        def gru         = Nadesico.conectar().getGrupo(ID)
-
-        formGrupo.setTxtIDGrupo        gru.id_grupo      as String
-        formGrupo.setTxtDescripcion    gru.descripcion
-        formGrupo.setTxtDescuento      gru.descuento     as String
-        formGrupo.setTxtUModificacion  gru.uModificacion as String
-        formGrupo.setModoDetalles();
-        formGrupo
+        if(cerrojo(PMA_DETALLESGRUPO)){
+            def formGrupo = new omoikane.formularios.Grupo()
+            formGrupo.setVisible(true)
+            escritorio.getPanelEscritorio().add(formGrupo)
+            Herramientas.iconificable(formGrupo)
+            formGrupo.toFront()
+            try { formGrupo.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario detalles Grupo", Herramientas.getStackTraceString(e)) }
+            def gru         = Nadesico.conectar().getGrupo(ID)
+            formGrupo.setTxtIDGrupo        gru.id_grupo      as String
+            formGrupo.setTxtDescripcion    gru.descripcion
+            formGrupo.setTxtDescuento      gru.descuento     as String
+            formGrupo.setTxtUModificacion  gru.uModificacion as String
+            formGrupo.setModoDetalles();
+            return formGrupo
+        }else{Dialogos.lanzarAlerta("Acceso Denegado")}
     }
 
     static def poblarGrupos(tablaMovs,txtBusqueda)
@@ -141,42 +133,31 @@ class Grupos {
 
     static def lanzarModificarGrupo(ID)
     {
-        
-        def formGrupo = new omoikane.formularios.Grupo()
-        formGrupo.setVisible(true)
-        escritorio.getPanelEscritorio().add(formGrupo)
-        Herramientas.iconificable(formGrupo)
-        formGrupo.toFront()
-        try { formGrupo.setSelected(true) } catch(Exception e) { Dialogos.lanzarDialogoError(null, "Error al iniciar formulario detalles Grupo", Herramientas.getStackTraceString(e)) }
-
-        def gru         = Nadesico.conectar().getGrupo(ID)
-
-        formGrupo.setTxtIDGrupo        gru.id_grupo      as String
-        formGrupo.setTxtDescripcion    gru.descripcion
-        formGrupo.setTxtDescuento      gru.descuento     as String
-        formGrupo.setTxtUModificacion  gru.uModificacion as String
+        def formGrupo = lanzarDetallesGrupo(ID)
         formGrupo.setModoModificar();
         formGrupo
-        
     }
 
     static def modificar(formGrupo)
     {
-        Herramientas.verificaCampos{
-        Herramientas.verificaCampo(formGrupo.getTxtDescripcion(),/^([a-zA-Z0-9_\-\s\ñ\Ñ\*\+áéíóúü]+)$/,"Descripcion sólo puede incluír números, letras, espacios, á, é, í, ó, ú, ü, _, -, * y +.")
-        Herramientas.verificaCampo(formGrupo.getTxtDescuento(),/^([0-9]*[\.]{0,1}[0-9]+)$/,"Descuento sólo puede incluír números reales positivos")
-
-        def serv = Nadesico.conectar()
-            Dialogos.lanzarAlerta(serv.modGrupo(formGrupo.getTxtIDGrupo(),formGrupo.getTxtDescripcion(),formGrupo.getTxtDescuento()))
-        }
+        if(cerrojo(PMA_MODIFICARGRUPO)){
+            Herramientas.verificaCampos{
+                Herramientas.verificaCampo(formGrupo.getTxtDescripcion(),/^([a-zA-Z0-9_\-\s\ñ\Ñ\*\+áéíóúü]+)$/,"Descripcion sólo puede incluír números, letras, espacios, á, é, í, ó, ú, ü, _, -, * y +.")
+                Herramientas.verificaCampo(formGrupo.getTxtDescuento(),/^([0-9]*[\.]{0,1}[0-9]+)$/,"Descuento sólo puede incluír números reales positivos")
+                def serv = Nadesico.conectar()
+                Dialogos.lanzarAlerta(serv.modGrupo(formGrupo.getTxtIDGrupo(),formGrupo.getTxtDescripcion(),formGrupo.getTxtDescuento()))
+            }
+        }else{Dialogos.lanzarAlerta("Acceso Denegado")}
     }
 
-        static def eliminarGrupo(ID)
+    static def eliminarGrupo(ID)
     {
-        def db = Sql.newInstance("jdbc:mysql://localhost/omoikane?user=root&password=", "root", "", "com.mysql.jdbc.Driver")
-        db.execute("DELETE FROM grupos WHERE id_grupo = " + ID)
-        db.close()
-        Dialogos.lanzarAlerta("Grupo " + ID + " supuestamente eliminado")
+        if(cerrojo(PMA_ELIMINARGRUPO)){
+            def db = Sql.newInstance("jdbc:mysql://localhost/omoikane?user=root&password=", "root", "", "com.mysql.jdbc.Driver")
+            db.execute("DELETE FROM grupos WHERE id_grupo = " + ID)
+            db.close()
+            Dialogos.lanzarAlerta("Grupo " + ID + " supuestamente eliminado")
+        }else{Dialogos.lanzarAlerta("Acceso Denegado")}
         
     }
 
@@ -186,4 +167,3 @@ class Grupos {
         reporte.lanzarPreview()
     }
 }
-
